Add files via upload
I cut the datasets into three:
The 2020 dataset starts in late January with 1 county and then grows as time continues into all counties. There will be missing rows until counties start showing cases.

The 2021 dataset shows data for all of 2021. This does have some unknown counties for a week or so in July.

The 20-21 dataset shows data from July 1st, 2020, to June 31st, 2021. There are no unknown rows, but there are some counties that did not push information on some days. I think the total rows that were not entered are around 40ish.